# Conflicts:
#	src/translations/en.json

+++ src/panels/config/person/dialog-person-detail.ts
@@ -1,4 +1,4 @@
-import { mdiPencil } from "@mdi/js";
+import { mdiAccount, mdiAccountPlus, mdiPencil } from "@mdi/js";
 import type { CSSResultGroup } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators";
@@ -19,6 +19,7 @@ import type { PersonMutableParams } from "../../../data/person";
 import type { User } from "../../../data/user";
 import {
   deleteUser,
+  fetchUsers,
   SYSTEM_GROUP_ID_ADMIN,
   SYSTEM_GROUP_ID_USER,
   updateUser,
@@ -36,6 +37,8 @@ import { documentationUrl } from "../../../util/documentation-url";
 import { showAddUserDialog } from "../users/show-dialog-add-user";
 import { showAdminChangePasswordDialog } from "../users/show-dialog-admin-change-password";
 import type { PersonDetailDialogParams } from "./show-dialog-person-detail";
+import { showListItemsDialog } from "../../../dialogs/dialog-list-items/show-list-items-dialog";
+import { computeDomain } from "../../../common/entity/compute_domain";
 
 const includeDomains = ["device_tracker"];
 
@@ -84,6 +87,8 @@ class DialogPersonDetail
   @state() private _personExists = false;
 
   @state() private _open = false;
+
+  private _linkedExistingUser = false;
 
   private _deviceTrackersAvailable = memoizeOne((hass) =>
     Object.keys(hass.states).some(
@@ -140,7 +145,7 @@ class DialogPersonDetail
   private _dialogClosed() {
     // If we do not have a person ID yet (= person creation dialog was just cancelled), but
     // we already created a user ID for it, delete it now to not have it "free floating".
-    if (!this._personExists && this._userId) {
+    if (!this._personExists && this._userId && !this._linkedExistingUser) {
       const callback = this._params?.refreshUsers;
       deleteUser(this.hass, this._userId).then(() => {
         callback?.();
@@ -160,9 +165,11 @@ class DialogPersonDetail
       <ha-dialog
         .open=${this._open}
         .preventScrimClose=${this.isDirtyState}
-        header-title=${this._params.entry
-          ? this._params.entry.name
-          : this.hass!.localize("ui.panel.config.person.detail.new_person")}
+        header-title=${
+          this._params.entry
+            ? this._params.entry.name
+            : this.hass!.localize("ui.panel.config.person.detail.new_person")
+        }
         @closed=${this._dialogClosed}
       >
         <div>
@@ -202,96 +209,107 @@ class DialogPersonDetail
               <ha-switch
                 slot="end"
                 @change=${this._allowLoginChanged}
-                ?disabled=${this._user &&
-                (this._user.id === this.hass.user?.id ||
-                  this._user.system_generated ||
-                  this._user.is_owner)}
+                ?disabled=${
+                  this._user &&
+                  (this._user.id === this.hass.user?.id ||
+                    this._user.system_generated ||
+                    this._user.is_owner)
+                }
                 .checked=${this._userId}
               ></ha-switch>
             </ha-row-item>
 
             ${this._renderUserFields()}
-            ${this._deviceTrackersAvailable(this.hass)
-              ? html`
-                  <p>
-                    ${this.hass.localize(
-                      "ui.panel.config.person.detail.device_tracker_intro"
-                    )}
-                  </p>
-                  <ha-entities-picker
-                    .hass=${this.hass}
-                    .value=${this._deviceTrackers}
-                    .includeDomains=${includeDomains}
-                    .pickedEntityLabel=${this.hass.localize(
-                      "ui.panel.config.person.detail.device_tracker_picked"
-                    )}
-                    .pickEntityLabel=${this.hass.localize(
-                      "ui.panel.config.person.detail.device_tracker_pick"
-                    )}
-                    @value-changed=${this._deviceTrackersChanged}
-                  >
-                  </ha-entities-picker>
-                `
-              : html`
-                  <p>
-                    ${this.hass!.localize(
-                      "ui.panel.config.person.detail.no_device_tracker_available_intro"
-                    )}
-                  </p>
-                  <ul>
-                    <li>
-                      <a
-                        href=${documentationUrl(
-                          this.hass,
-                          "/integrations/#presence-detection"
-                        )}
-                        target="_blank"
-                        rel="noreferrer"
-                        >${this.hass!.localize(
-                          "ui.panel.config.person.detail.link_presence_detection_integrations"
-                        )}</a
-                      >
-                    </li>
-                    <li>
-                      <a @click=${this.closeDialog} href="/config/integrations">
-                        ${this.hass!.localize(
-                          "ui.panel.config.person.detail.link_integrations_page"
-                        )}</a
-                      >
-                    </li>
-                  </ul>
-                `}
+            ${
+              this._deviceTrackersAvailable(this.hass)
+                ? html`
+                    <p>
+                      ${this.hass.localize(
+                        "ui.panel.config.person.detail.device_tracker_intro"
+                      )}
+                    </p>
+                    <ha-entities-picker
+                      .value=${this._deviceTrackers}
+                      .includeDomains=${includeDomains}
+                      .pickedEntityLabel=${this.hass.localize(
+                        "ui.panel.config.person.detail.device_tracker_picked"
+                      )}
+                      .pickEntityLabel=${this.hass.localize(
+                        "ui.panel.config.person.detail.device_tracker_pick"
+                      )}
+                      @value-changed=${this._deviceTrackersChanged}
+                    >
+                    </ha-entities-picker>
+                  `
+                : html`
+                    <p>
+                      ${this.hass!.localize(
+                        "ui.panel.config.person.detail.no_device_tracker_available_intro"
+                      )}
+                    </p>
+                    <ul>
+                      <li>
+                        <a
+                          href=${documentationUrl(
+                            this.hass,
+                            "/integrations/#presence-detection"
+                          )}
+                          target="_blank"
+                          rel="noreferrer"
+                          >${this.hass!.localize(
+                            "ui.panel.config.person.detail.link_presence_detection_integrations"
+                          )}</a
+                        >
+                      </li>
+                      <li>
+                        <a
+                          @click=${this.closeDialog}
+                          href="/config/integrations"
+                        >
+                          ${this.hass!.localize(
+                            "ui.panel.config.person.detail.link_integrations_page"
+                          )}</a
+                        >
+                      </li>
+                    </ul>
+                  `
+            }
           </div>
         </div>
         <ha-dialog-footer slot="footer">
-          ${this._params.entry
-            ? html`
-                <ha-button
+          ${
+            this._params.entry
+              ? html`
+                  <ha-button
+                    slot="secondaryAction"
+                    variant="danger"
+                    appearance="plain"
+                    @click=${this._deleteEntry}
+                    .disabled=${
+                      (this._user && this._user.is_owner) || this._submitting
+                    }
+                  >
+                    ${this.hass!.localize("ui.panel.config.person.detail.delete")}
+                  </ha-button>
+                `
+              : html`<ha-button
                   slot="secondaryAction"
-                  variant="danger"
                   appearance="plain"
-                  @click=${this._deleteEntry}
-                  .disabled=${(this._user && this._user.is_owner) ||
-                  this._submitting}
+                  @click=${this.closeDialog}
                 >
-                  ${this.hass!.localize("ui.panel.config.person.detail.delete")}
-                </ha-button>
-              `
-            : html`<ha-button
-                slot="secondaryAction"
-                appearance="plain"
-                @click=${this.closeDialog}
-              >
-                ${this.hass!.localize("ui.common.cancel")}
-              </ha-button>`}
+                  ${this.hass!.localize("ui.common.cancel")}
+                </ha-button>`
+          }
           <ha-button
             slot="primaryAction"
             @click=${this._updateEntry}
             .disabled=${nameInvalid || this._submitting || !this.isDirtyState}
           >
-            ${this._params.entry
-              ? this.hass!.localize("ui.common.save")
-              : this.hass!.localize("ui.common.add")}
+            ${
+              this._params.entry
+                ? this.hass!.localize("ui.common.save")
+                : this.hass!.localize("ui.common.add")
+            }
           </ha-button>
         </ha-dialog-footer>
       </ha-dialog>
@@ -302,56 +320,64 @@ class DialogPersonDetail
     const user = this._user;
     if (!user) return nothing;
     return html`
-      ${!user.system_generated
-        ? html`
-            <ha-row-item>
-              <span slot="headline"
-                >${this.hass.localize(
-                  "ui.panel.config.person.detail.username"
-                )}</span
-              >
-              <span slot="supporting-text">${user.username}</span>
-              ${this.hass.user?.is_owner
-                ? html`
-                    <ha-icon-button
-                      slot="end"
-                      .path=${mdiPencil}
-                      @click=${this._changeUsername}
-                      .label=${this.hass.localize(
-                        "ui.panel.config.person.detail.change_username"
-                      )}
-                    >
-                    </ha-icon-button>
-                  `
-                : nothing}
-            </ha-row-item>
-          `
-        : nothing}
-      ${!user.system_generated && this.hass.user?.is_owner
-        ? html`
-            <ha-row-item>
-              <span slot="headline"
-                >${this.hass.localize(
-                  "ui.panel.config.person.detail.password"
-                )}</span
-              >
-              <span slot="supporting-text">************</span>
-              ${this.hass.user?.is_owner
-                ? html`
-                    <ha-icon-button
-                      slot="end"
-                      .path=${mdiPencil}
-                      @click=${this._changePassword}
-                      .label=${this.hass.localize(
-                        "ui.panel.config.person.detail.change_password"
-                      )}
-                    >
-                    </ha-icon-button>
-                  `
-                : nothing}
-            </ha-row-item>
-          `
-        : nothing}
+      ${
+        !user.system_generated
+          ? html`
+              <ha-row-item>
+                <span slot="headline"
+                  >${this.hass.localize(
+                    "ui.panel.config.person.detail.username"
+                  )}</span
+                >
+                <span slot="supporting-text">${user.username}</span>
+                ${
+                  this.hass.user?.is_owner
+                    ? html`
+                        <ha-icon-button
+                          slot="end"
+                          .path=${mdiPencil}
+                          @click=${this._changeUsername}
+                          .label=${this.hass.localize(
+                            "ui.panel.config.person.detail.change_username"
+                          )}
+                        >
+                        </ha-icon-button>
+                      `
+                    : nothing
+                }
+              </ha-row-item>
+            `
+          : nothing
+      }
+      ${
+        !user.system_generated && this.hass.user?.is_owner
+          ? html`
+              <ha-row-item>
+                <span slot="headline"
+                  >${this.hass.localize(
+                    "ui.panel.config.person.detail.password"
+                  )}</span
+                >
+                <span slot="supporting-text">************</span>
+                ${
+                  this.hass.user?.is_owner
+                    ? html`
+                        <ha-icon-button
+                          slot="end"
+                          .path=${mdiPencil}
+                          @click=${this._changePassword}
+                          .label=${this.hass.localize(
+                            "ui.panel.config.person.detail.change_password"
+                          )}
+                        >
+                        </ha-icon-button>
+                      `
+                    : nothing
+                }
+              </ha-row-item>
+            `
+          : nothing
+      }
       <ha-row-item>
         <span slot="headline"
           >${this.hass.localize(
@@ -405,26 +431,71 @@ class DialogPersonDetail
     this._updateDirtyState(this._currentState());
   }
 
+  private async _linkUser(user: User, newUser: boolean) {
+    this._linkedExistingUser = !newUser;
+    if (this._params!.entry && this._params!.updateEntry) {
+      await this._params!.updateEntry({ user_id: user.id });
+    }
+    if (newUser) {
+      this._params?.refreshUsers?.();
+    }
+    this._user = user;
+    this._userId = user.id;
+    this._isAdmin = user.group_ids.includes(SYSTEM_GROUP_ID_ADMIN);
+    this._localOnly = user.local_only;
+    this._updateDirtyState(this._currentState());
+  }
+
   private async _allowLoginChanged(ev): Promise<void> {
     const target = ev.target;
     if (target.checked) {
       target.checked = false;
-      showAddUserDialog(this, {
-        userAddedCallback: async (user?: User) => {
-          if (user) {
-            target.checked = true;
-            if (this._params!.entry && this._params!.updateEntry) {
-              await this._params!.updateEntry({ user_id: user.id });
+
+      const users = await fetchUsers(this.hass);
+      const currentLinkedUsers = new Set(
+        Object.values(this.hass.states)
+          .filter(
+            (s) =>
+              computeDomain(s.entity_id) === "person" && s.attributes.user_id
+          )
+          .map((s) => s.attributes.user_id)
+      );
+      const eligibleUsers = users.filter(
+        (u) =>
+          !currentLinkedUsers.has(u.id) && !u.system_generated && u.username
+      );
+      const addUserDialog = () =>
+        showAddUserDialog(this, {
+          userAddedCallback: async (user?: User) => {
+            if (user) {
+              target.checked = true;
+              this._linkUser(user, true);
             }
-            this._params?.refreshUsers?.();
-            this._user = user;
-            this._userId = user.id;
-            this._isAdmin = user.group_ids.includes(SYSTEM_GROUP_ID_ADMIN);
-            this._localOnly = user.local_only;
-            this._updateDirtyState(this._currentState());
-          }
-        },
-        name: this._name,
+          },
+          name: this._name,
+        });
+
+      if (eligibleUsers.length === 0) {
+        addUserDialog();
+        return;
+      }
+
+      showListItemsDialog(this, {
+        title: this.hass.localize("ui.panel.config.person.detail.select_user"),
+        items: [
+          {
+            iconPath: mdiAccountPlus,
+            label: this.hass.localize(
+              "ui.panel.config.person.detail.create_new_user"
+            ),
+            action: addUserDialog,
+          },
+          ...eligibleUsers.map((user) => ({
+            iconPath: mdiAccount,
+            label: `${user.name} (${user.username})`,
+            action: () => this._linkUser(user, false),
+          })),
+        ],
       });
     } else if (this._userId) {
       if (
